+++ src/RuleCollection.php
@@ -36,6 +36,7 @@
 use Glpi\Application\View\TemplateRenderer;
 use Glpi\DBAL\QueryExpression;
 use Glpi\Event;
+use Glpi\Plugin\Hooks;
 
 class RuleCollection extends CommonDBTM
 {
@@ -1629,15 +1630,15 @@ TWIG, $twig_params);
         global $PLUGIN_HOOKS;
 
         $input = $this->prepareInputDataForProcess($input, $params);
-        if (isset($PLUGIN_HOOKS['use_rules'])) {
-            foreach ($PLUGIN_HOOKS['use_rules'] as $plugin => $val) {
+        if (isset($PLUGIN_HOOKS[Hooks::USE_RULES])) {
+            foreach ($PLUGIN_HOOKS[Hooks::USE_RULES] as $plugin => $val) {
                 if (!Plugin::isPluginActive($plugin)) {
                     continue;
                 }
                 if (is_array($val) && in_array(static::getRuleClassName(), $val, true)) {
                     $results = Plugin::doOneHook(
                         $plugin,
-                        'ruleCollectionPrepareInputDataForProcess',
+                        Hooks::AUTO_RULE_COLLECTION_PREPARE_INPUT_DATA_FOR_PROCESS,
                         ['rule_itemtype' => static::getRuleClassName(),
                             'values'        => ['input' => $input,
                                 'params' => $params,
@@ -1802,16 +1803,16 @@ TWIG, $twig_params);
         /** @var array $PLUGIN_HOOKS */
         global $PLUGIN_HOOKS;
 
-        if (isset($PLUGIN_HOOKS['use_rules'])) {
+        if (isset($PLUGIN_HOOKS[Hooks::USE_RULES])) {
             $params['rule_itemtype'] = static::class;
-            foreach ($PLUGIN_HOOKS['use_rules'] as $plugin => $val) {
+            foreach ($PLUGIN_HOOKS[Hooks::USE_RULES] as $plugin => $val) {
                 if (!Plugin::isPluginActive($plugin)) {
                     continue;
                 }
                 if (is_array($val) && in_array($params['rule_itemtype'], $val, true)) {
                     $results = Plugin::doOneHook(
                         $plugin,
-                        "preProcessRuleCollectionPreviewResults",
+                        Hooks::AUTO_PRE_PROCESS_RULE_COLLECTION_PREVIEW_RESULTS,
                         ['output' => $output,
                             'params' => $params,
                         ]